+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- PostgreSQL Cmd+K database switcher now lists and switches databases and schemas for postgres 
+
 ## [0.13.0] - 2026-03-04
 
 ### Added